backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# LangChain & LLM
langchain==0.1.6
langchain-google-genai==0.0.11
langchain-community==0.0.19

# Vector Store
faiss-cpu==1.7.4
# chromadb==0.4.22

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
tabulate==0.9.0
pandas==2.1.4

# Embeddings
sentence-transformers==2.3.1

# Validation & Data
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1